+++ domain/src/main/java/it/to/peppesca/domain/models/PokemonEntity.kt
@@ -3,4 +3,4 @@ package it.to.peppesca.domain.models
 /**
  * Entity model for domain layer
  */
-data class PokemonEntity(val id: Long, val name: String)
+data class PokemonEntity(val id: Int, val name: String)
